Removed duplicate unit tests
